+++ CHANGELOG.md
@@ -8,7 +8,7 @@ All notable changes to this project are documented here.
 - Moved `bbstat.evaluate.BootstrapResult.ndigits` to `bbstat.utils.get_precision_from_credibility_interval`.
 - Moved `bbstat.evaluate.credibility_interval` to `bbstat.utils.compute_credibility_interval`.
 - Moved `bbstat.evaluate.BootstrapResult.plot` to `bbstat.plot.plot`.
-- Renamed all instances "credibility" -> "credible".
+- Renamed all instances "credibility" -> "credible" and "coverage" -> "level".
 
 ## [0.1.0] - 2025-10-27
 Core logic and selected statistic functions.